BETA 3 commit
This is not the BETA 3 announced before. I decided
to name this beta 3 to avoid problems to release so a big
change as beta 2, the second release.
The new beta number will allow me to create new packages
which are needed after all that changes.

This is commit will fix some small problems and
- new wilderness maps (rusty)
- new underground dungeon (cat)
- dark temple map (NOT COMMITED HERE!)
- dozen new arches
- dozens new monsters
- reworked exp/level/mob stat system
- rework color target table
- alot more small fixes