artifact patch - speed up artifact list finder by using type based table - added hash table for artifacts like for arches - added find_artifact (mask) using hash table - added "amask" object command to allow   explicit artification when loaded - added CreateArtifact" for scripts to allow   explicit artification in scripts
Very useful stuff - the artifacts and artifact masking (artification)
is now usable on the same level as archetype functions. The amask
object command allows to assign artification of objects inside the maps
with allowing global changes in artifacts file because "late binding"
of the mask data in the load_objects() of the server. TODO: editor
needs an interface to select and apply (and perhaps show) this kind
of data.