Re-introducing rings, amulets and magic devices to the treasurelist. The main treasurelist for all this is loot_magic. This should be specified in the randomitems attribute in mobs. To get roughly the same chances of everything dropping like they did in beta 3, set it to around &d10 for a spellcaster, and &d27 for a melee fighter. Some weaker mobs such as kobolds and slugs had a much lower chance than the others, and to get roughly the same chance with them means setting it to around &d42.
Since this hasn't been done yet, this treasurelist isn't activated yet. It also has a side-effect of temporarily removing the potion of improvement from what was activated, but seeing as it had around a 1/50000 chance of dropping, it wouldn't make much difference. :P

Oh, and also re-introduced traps, which are now activated in all corpses (apart from diseased needles) so beware! :)

And least importantly, there are now no more magic torches dropped, and the number of torch drops are cut down a bit too.