Fix to a change I made a while ago.
I made decline append #n to it's command where an icon is selected (as does accept).

However this is no good when you specifically want the button (ie, decline for current interfaces or 'accept' (ie, non-action) for SENTInce) not to have its command modified. That is, the button is declining all choices.

Therefore, button commands need to explicitly indicate if they're going to allow modification by being prefixed with #. Eg, #accept beomes accept #1 while decline is only ever decline. This makes more sense and is not just an unnecessary complexity where talk commands are 'built up'. For example, 'buy' opens a list of goods for sale. The player selects one, clicks a button (command '#buy') and the processed command becomes 'buy #n', or clicks the other button (command 'buy') and the ptocessed command is just 'buy' (ie, reopen the list).