church.lua -- EXAMPLE ONLY
A generic priest script. Due to the lack of gods in Daimonin currently, only the Tabernacle, Moroch, and none are currently recognised by the script.
The priest object should have its script data set to the name of the god and the script kind of expects there to be an altar to the god in the church.
This is a generic script for demo purposes and should not be use on a public server.