* Amended a couple of wrong calculations. [interface.c] * Removed some unnecessary blank lines. [interface.c] * Changed position of quantity text in icons from brc to blc. This means quantities from -999 to +999 display well. Quantities outside of that range overflow the icon box which looks a bit crapp, but there's not much to be done about that. That range should be sufficient (icons are not intended for coins, which are displayed separately). [interface.c] * Added alternatives ib:DeactivateIcons() and ib:ActivateIcons() for ib:SelectOff() and ib:SelectOn(), and adjusted the (internal) working of ib.select.mode. [interface_builder.lua] * Added handling of icon mode 'g'. [interface.c and interface_builder.lua] * Changed icon outline box to BITMAP_INVSLOT_U for unselected/deactivated icons and BITMAP_INVSLOT for selected/activated icons. [interface.c]
All this means that in NPC GUIs icon background colour indicates purpose of icon:
  * red = icon represents object(s) removed from player's inv (signed quantity shown in red in corner of icon);
  * green = icon represents object(s) added to player's inv (signed quantity shown in green in corner of icon);
  * no fill = icon represents non-inv object/information only (such as guild membership, spell, or skill), no quantity shown;
  * grey = icons -- all icons, not just selectable ones -- in preview mode/deactivated (signed quantity, if relevant shown in red or green in corner of icon).

And the outline box indicates whether that icon is active/selected or not.

The 'drawback' of this is that it is not compatible with SENTInce-unaware clients. This should not be a problem if we get the autoupdater. But until then, this feature is turned OFF in interface_builder.lua (which is done by never sending <im="g"> to the client, just <im="G"> instead). Meaning that non-selectable icons always appear active, as usual.

Finally, this has left some duplicate code in interface.c ripe for tidying, which I'll do later.