The server shows a 'MOTD' (data/motd) to new characters (but not existing characters). This MOTD is also accessible to all players with the command /motd. But it is read-only.
This patch shows the MOTD to all players on login (or via /motd). DMs can also write a new MOTD with /motd <message>.

There are some extra features which I am too lazy to add (exercise for the reader ;)):

The default MOTD currently cannot be restored. The default file is always untouched but there si yet no way to delete the written file. Simple job. Make it so.

The default MOTD should be always the default, but the written MOTD should have a (optional) timeout. Fairly simple job. Make it so.