The level of a gate is settable.
The gate's level determines how hard it hits a damageable object that is crushed by it closing.

If level is 0 (the default), the value will be calculated according to map difficulty.

This means players/monsters that get caught in a gate in a level 100 dungeon take reasonable damage (301)
rather than the old 4. But a gate in a level 1 dungeon only does 4 damage.
In any case the gate's level may be explicitly set for flexibility.
Code patch posted to forum due to state of trunk/server (So ATM this won't have any effect).