Pedestals in Trigger pedestals have 'match = having' and 'blocking passage' flags like Inv Checkers.
Pedestals, Trigger Pedestals, and Inv Checkers have a message filed so they optionally print a message to the player that triggered them on a match.

Code patch posted to forum due to state of trunk/server (So ATM this won't have any effect).

TODO: Add a level req field to pedestals, which was the whole point of starting this. I forgot...