Added a level req to pedestals, and more control over messaging and planned out their activation a bit better (note that temporarily inv checkers may not, in fact will not, message properly).
Here's an overview of pedestal bebhaviour:

THE TWO USES OF PEDESTALS
~~~~~~~~~~~~~~~~~~~~~~~~~
Pedestals are activated by players or monsters trying to enter or entering their square.
Pedestals can either restrict passage (block a square) or trigger a connection (but not both).

SUCCESS AND FAILURE IN MATCHING
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
'Match text' (slaying) is a string which is matched against the activator's race,
or it may be 'player', a special case which matches all players.

Pedestals may be 'match = having' or not (last_sp = 1 or 0), meaning that the match is
a success if the 'match text' does match the activator (1) or if it does not (0).
For example, 'slaying = "elf"; last_sp = 1' means all elves will match,
'slaying = "elf"; last_sp = 0' means anything but elves will match.  

If we so far have a match, the activator must pass an optional level test ('level requirement', level).
If level = 0, there is no level requirement.
If level > 0, activator.level must be >= pedestal.level.
If level < 0, activator.level must be <= 0 - pedestal.level.

If we pass the level requirement too, or it doesn't apply, the match is a success.

BLOCKING PASSAGE PEDESTALS
~~~~~~~~~~~~~~~~~~~~~~~~~~
For 'blocking passage' (last_grace = 1) pedestals, only activators successfully matching can enter the square.
While the pedestal may have a 'connection' (connected) value, this will be ignored.

CONNECTED PEDESTALS
~~~~~~~~~~~~~~~~~~~
For not 'blocking passage' (last_grace = 0) pedestals, anyone may enter the square
but only activators successfully matching will trigger the connection.

MESSAGING
~~~~~~~~~
If a pedestal has a 'message' (msg), this text is shown to the activator (players only of course) of the pedestal.
'Message all nearby' (item_condition = 1) pedestals also show 'message' to all players nearby.

For 'message on success' (item_quality = 1) pedestals, the message is shown only if the match is a success,
ie, when the square is entered/connection triggered, depending on last_grace.

For not 'message on success' (item_quality = 0) pedestals, the message is only shown if the match is not a success,
ie, when the square is not enterable/the connection is not triggered by the activator, depending on last_grace.