Gecko wrote: > Please avoid both complex assignments and tests in the same line. Same for > returns: > ... > And even more important: > this function is called "check_button_down" which implies that it doesn't > have any side effects. So either rename the function or remove the > op->weight_limit = 1; > > If this function is only used from this file, make it static to indicate > that it is a local utility function.
I fixed the first issue. Thanks.

As to the second one, while I very much agree on all points and had planned to make such changes later on (that is if I don't make a more substantial change which will make this issue irrelevant), to be fair to me, this is the existing trunk code, so not my fault, guv.