Candlelights become animated applyable lights rather than static misc objects.
This means players can light and extinguish them. Like wall torches they are
permanent.

Note that they are lit by default, to fit better with existing practices (also
it would be daft to have a map plunged into darkness as soon as it loads or is
reset) but there seems to be a buglet in the server so they show the unlit face
until you extinguish and relight them.