Deliver 'gloomy' stream (r3276-3277):
Currently maps either have a fixed darkness (brightness) or they are flagged as
outdoors meaning the ambient light level varies over the course of a day. This
means that all outdoors maps (ie, on all planes) have the same light levels.

This patch means that you can now give an outdoors map a darkness value as
well, and the variable ambient light will then be capped at that level.
Meaning that we can, for example, give all outdoors demon plane maps a darkness
of, say, 4 for an overall 'gloomy' effect.

By default, the server gives maps a darkness of -1. Until now there has been no
reason to set darkness on outdoors maps, so mapmaker's haven't. Thus the
default value is used. This means that the variable light will be *unaffected*
by this patch on these maps, which is a good thing. So it is (should be)
completely backwards compatible.