Merging op->action_timer into op->ob->weapon_speed_left. There's still a couple more odds and ends to do to finish this.
Also, ranged weapons are no longer restricted to having a delay time in increments of 0.125 seconds, which is a nice benefit of op->ob->weapon_speed_left. As a result, any ranged weapon arches/artifacts will need their last_grace attribute changed. This shouldn't be a problem because all artifacts on the main server are referenced directly so there shouldn't be any ranged weapons in the game with the attribute hard written to the base arch.

Finally, it also means check_skill_action_time() is no longer used for melee combat, otherwise it would mean skill_attack() gets called every tick if a player has a mob targetted. It doesn't matter too much before much of the code in there will be removed shortly.