Make the level of a gate settable.
The gate's level determines how hard it hits a damageable (live, player or mob)
object that is crushed by it closing.

If level is 0 (the default), the value will be calculated according to map
difficulty.

This means players/monsters that get caught in a gate in a level 100 dungeon
take reasonable damage (301) rather than the old 4. But a gate in a level 1
dungeon only does 4 damage. This is for players of the same level as the
dungeon/gate. For other levels, the damage is calculated so that, eg, a level 1
will still be creamed by a level 100 gate (202 dam), but a level 100 might
think twice about standing under a level 1 gate (301 dam). In any case the
gate's level may be explicitly set for flexibility.