Removing the "You can shoot again in x seconds." spam. Now we have a new variable that is displayed in the client for that job. The server now idealy only updates the client twice whenever a player uses a skill. The first to send the initial delay to the client, and the second to tell the client that the cooldown time is over.
There are still several things left to do. The client needs to tick this counter down pre-emptively. The new variable is currently displayed on the skill exp widget but it should preferably be in its own widget. Finally, it's now possible to make the client block skill using signals to the server if the counter is greater than zero, reducing more bandwidth.

This commit means the server in this stream is now no longer compatible with the trunk client, as a new variable is sent to the client. The client from this stream must be used with the server from this stream now.