Fix a bug with the new method of displaying skill delay. It's possible that you use a skill exactly when the cooldown time hits 0. If that happens, the server would send the client the new skill delay time without telling the client that the cooldown time has ended. If the new skill delay time is exactly the same as the last one, the server would think the value is unchanged and fail to update the client. While this would usually not be a problem, the client will pre-emptively tick this down, and so will not show the new skill delay.
So here we use a little trick. If the last value sent to the client was positive, we change the sign to negative so that the server will send the new value always when a skill delay is set. Not exactly the tidiest of solutions, but it does the job well.