Adding the ability to have multiple widgets at the same time. Right now, this is only partially implemented for text windows, but it shouldn't be too much of a job to convert the other widgets too. This is done by adding next and prev pointers in the _widgetdata struct, and changing the cur_widget _widgetdata array into an array of pointers to _widgetdata, effectively creating an array of linked lists.
The widgets should behave perfectly normal atm, since most of them are only occupying the head node of each linked list, and even the text windows occupying the head node should be fine too. The partial implementation is only for creating more text windows.

In order to create new widgets, I've hacked the macro code for now, though eventually there should be buttons in order to do this. ?M_CREATEWIDGET will create a new CHATWIN_ID widget, which is added to the end of the linked list, and ?M_REMOVEWIDGET will remove a CHATWIN_ID widget with a WidgetObjID of 3. It's pretty random, but it's pre-alpha code so expect a bit of weird stuff. Also, this is *not* the same as removing node #3, and there's a reason for that. Also I've yet to make a function to remove a widget if it's simply handed the pointer, which would mean no traversing is needed if removed that way. I could do with a better command for creating/removing widgets for now though until proper buttons are made.

There is a lot of debug output atm, since the linked list code is brand new, but I've given it a test and it seems to work. Anyway, this stream right now is highly experimental, at least until all the widgets make full use of the linked lists.