Ranged attacks now use the new damage system. I've also made some adjustments to it, so the dps value shown in the client is currently incorrect, but apparently this value has always been wrong. The code here to calculate damage should be in one place to prevent these kinds of problems, so I'll sort that out next.
The adjustments to the ranged attack system mean these changes. Arrows no longer add to the skill delay. Only bows give a delay now. This is done to simplify the system and prevent an issue where faster but less powerful arrows can be better than more damaging but slower arrows. Arrows also no longer add to the base damage of a bow. It instead adds to the damage bonus, which is added to that of rings, etc. Then the base damage is multiplied by this damage bonus.