Replace 'name name§media' with 'name name', 'background_music media'.
ATM the client/server can handle both old and new style for backwards
compatibility. This is the tiniest bit more work than is needed so should
eventually be removed (as noted in the source) -- ie, when we have the latest
Gridarta in our MapMaker so mappers are producing new-style maps only).

Also I have disabled the client's ability to use PNG media, as this seems
pointless nowadays and is unused anyway: so media /only/ means music.