Commit some code that was lost when the hard drive died back at Christmas. This sees the beginnings of submenus implemented.
Be warned, this commit contains a major bug that creates a new instance of a context menu with every frame whenever a context menu is open. Closing the context menu will cause all instances to be removed from memory, however.

Normally I wouldn't commit anything with a bug as serious as this, but if it doesn't land in the SVN I won't get any piece of mind, given the fact that this code has been inaccessible for months.