Adding a potential replacement for the current image of the chain legs. I thought I committed this ages ago, but I guess it managed to sneak away.
Anyway, a pet peeve I have with the current images of pants/legs is that they look awfully rigid and ugly. They also seem to have an awkwardly wide gap between the legs. I know the original images weren't planned to be pants, but there have been images added later that seem based on this rigid design.

The replacement image isn't perfect either, as I know there has been criticism regarding the angle, which is a by-product of some experimentation that didn't really work out. However, I think it is still light-years ahead of the image it is replacing. That said, I may redo the image in future with a more normal angle.

I know this is a stupidly long commit message for such a small addition, but I hope I got my message across that pants should not look square and rigid. Rant over. :)