Add two scripts from MapWiz that should be public.
insert_marker.lua is a script to insert a force in a player like a regular marker does, except that it can be done without having to walk over a marker.

remote_connection_timed.lua is similar to remote_connection.lua except it behaves like the handle_trigger counterpart to the handle. It might be possible to merge these two scripts together by making the third parameter optional.