Fixed a minor memory leak in mkdir_recurse (memory allocated using strdup should be freed).
For windows OS (later version than Win98), a Daimonin directory is created in the %APPDIR% directory and the client behaves as in Unix (some files are copied from the Daimonin directory).

For Win98 OS, when %APPDIR% is not defined, the %WINDIR%\Application Data\Daimonin\ directory is created (may not be the better choice???) and the client behaves as in Unix (some files are copied from the Daimonin directory).

Current status of tests:
Implemented and tested under Windows XP.
Not tested under Vista (I will do so soon).
Not tested under Win98 (I will do so soon).
Not tested under Windows 2000 (won't be able to do so - but should work as for Windows XP).
Not tested under Windows 7 (won't be able to do so).
Non regression testing not done under Unix (will do so soon).