Fix artifact initialisation/generation.
Explanation: In r1281, the cp+=7 was added. This is wrong, it should be cp+=8
('Allowed '). So the continue clauses for all and none (and therefore the
setting of none_flag) are never actually reached. Thus all and none are added
to art->allowed->name just like a real arch ID. So the artifact system has
been broken for a very long time.

r2534 then saw fit to change a check for art->allowed == NULL /* ie all */ to
!strcmp(art->allowed->name, "all"), putting the seal of approval on this broken
behaviour.