Add ability to plot other players on the locator.
The way this works is that a client sends its geolocation (which is currently
obtained via hostip.info) in the setup string. This is then written to the
player's socket struct. Subsequent pings get the geolocation for all
non-privacy seeking, online players sent in return. The client can then plot on
the locator.