Player CLF objects are now stored in inventory of a CLAN_CONTAINER object (similar to quests and quest-containers).  This gets rid of our linked list of CLF objects using next/prev, and instead allows a more standard above/below linked list, as per most other inv. items.
Note - this also paves the way for allowing ranking (using prev/next) of a player's own CLF objects.

Also MOTD functionality reinstated, as it was accidently deleted in a previous commit.

Minor bug-fix when player logs on and has been given a different role in a clan (his old role was displayed, instead of his new role).