Lot of changes here!  Updates in lots of different areas required to introduce a 4th map type (currently we have multi / instance / unique), and now we add 'clan'.
This is required to support clan specific apartments.  Clan maps are basically the same as unique maps, but are stored in the /clans folders rather than the /players folder.

Basic testing done only so far.  This code won't really 'run' without suitable NPC scripts, which aren't included in this commit.