Make server update clients a necessary when an object is inserted in a player-viewed environment.
Note that currently the old code contains many specialised (and often
incomplete) versions of this, meaning that frequently player-initiated
actions which result in an object insertion, such as getting, buying, etc --
will send duplicate ITEM cmds to the client. This should however be
unnoticeable to players and anyway will be fixed in the next(ish) commits.